% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{animate_close}
\alias{animate_close}
\title{Animates the simulated closing of a jaw}
\usage{
animate_close(
  x,
  addjaw = FALSE,
  jawsp = NULL,
  render = TRUE,
  fps = 10,
  shadow = FALSE
)
}
\arguments{
\item{x}{a \code{list} returned by \code{close_jaw()}.}

\item{addjaw}{logical should a cartoon of a jaw and head be plotted}

\item{jawsp}{character, head and jaw species. See \code{details}.}

\item{render}{should plot be rendered with \code{gganimate::animate()}. See \code{details}.}

\item{fps}{numeric, frames per second controlling playback speed of animation.}

\item{shadow}{logical, should shadow of previous positions be plotted.}
}
\value{
a gif \code{magick-image} object if \code{render=TRUE}, a \code{gganim} object if \code{render=FALSE}. See \code{details}.
}
\description{
Plots the output of important positions established simulated by \code{close_jaw()} throughout adduction.
}
\details{
This function returns a plot of the origin and insertions positions as well as the jaw based on the output of data passed through \code{start_sim}.

Plotted data include positions origins, insertions, and lines of action/muscle lengths for the adductor muscle divisions, the Meckelian tendon, and jaw. If \code{addjaw=TRUE} a cartoon of  the head and jaw of a limited number of species indicated by \code{jawsp} is also plotted beneath these data. Current species option include a viperfish (\code{jawsp="Viperfish"}), a largemouth bass (\code{jawp="LMBass"}), wrasse (\code{jawsp="Wrasse"}), or gar (\code{jawsp="Gar"}).

If \code{render=TRUE}, a \code{magick-image} with playback speed and a shadow determined by the arguments \code{fps} and \code{shadow}. Users may want more flexibility in controlling \code{gganimate} options and therefore may use \code{render=FALSE} to return a \code{gganim} object, add options and pass it to \code{gganimate:animate()}. The data passed from \code{x} are identical in both cases.
}
\examples{

#without a jaw
start_sim(config=3) \%>\%
   close_jaw() \%>\%
    {. ->> cs } \%>\%
   animate_close(addjaw=TRUE,jawsp="LMBass")

#with a jaw (oddly, a wrasse jaw)

cs \%>\%
   animate_close(addjaw=TRUE,jawsp="LMBass")

cs_an <- cs \%>\%
   animate_close(render=FALSE)

#looks terrible
cs_an2 <- cs_an+shadow_wake(0.1, size = 2, alpha = FALSE, colour = 'grey92')
animate(cs_an2)

x=cs
}
