% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2.R
\name{close_jaw}
\alias{close_jaw}
\title{Jaw closing simulations}
\usage{
close_jaw(x, MaxIts = 1000, progress = F)
}
\arguments{
\item{x}{a list obtained by \code{start_sim}}

\item{MaxIts}{numeric, how long should the simulation proceed.}

\item{progress}{logical, .....}
}
\description{
Establishes starting values for important morphological and physiological parameters to be used in adduction simulations with \code{close_jaw()}.
}
\details{
This function takes morphometric data of the feeding system and prepares it for analysis with \code{close_jaw}. It essentially uses the input data to open the jaw to a specified or default gape angle and outputs important muscle and model parameters to be used by \code{close_jaw}.

\code{x} should be a \code{data.frame} or  containing the the variables specified by \code{jaw_vars}.
}
\examples{

cs_dat<-read.csv(system.file("extdata", "csloani.csv", package = "JawCloseR")
cs_dat \%>\% start_sim()
sm <- start_sim(config=1,pars=list(Pressure=100)) \%>\% close_jaw()
#print(sm$ani)
sm$j \%>\% select(Time,DragT,PressureT) \%>\% pivot_longer(-Time) \%>\% ggplot(aes(Time,value))+geom_point()+facet_wrap(~name)
}
\seealso{
\code{\link{close_jaw}}
}
