% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{cos_ang}
\alias{cos_ang}
\title{Compute angle between two segments of a triangle using law of cosines.}
\usage{
cos_ang(l, r, o)
}
\arguments{
\item{l}{numeric,  length of segment to the left.}

\item{r}{length of segment to the right.}

\item{o}{length of opposite segment.}
}
\value{
A single value of the angle in radians.
}
\description{
Computes angle between two segments of a triangle using law of cosines.
}
