% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{cos_side}
\alias{cos_side}
\title{Compute length of as segment using law of
#' @description  Computes length of a segment given the length of two adjacent segments and their angle using law of cosines.}
\usage{
cos_side(l, r, ang)
}
\arguments{
\item{l}{numeric,  length of segment to the left.}

\item{r}{length of segment to the right.}

\item{and}{numeric, the angle shared by the segments.}
}
\value{
A single value the length.
}
\description{
Compute length of as segment using law of
#' @description  Computes length of a segment given the length of two adjacent segments and their angle using law of cosines.
}
