% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{dist_2d}
\alias{dist_2d}
\title{Compute distance between two points in Cartesian space}
\usage{
dist_2d(pt1, pt2)
}
\arguments{
\item{pt1}{numeric, vector of length 2 representing x,y position of the first point.}

\item{pt2}{numeric, vector of length 2 representing x,y position of the second point.}
}
\value{
A single value of the length.
}
\description{
Computes distance between two points in Cartesian space using simple trigonometry functions
}
