% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2.R
\name{jaw_vars}
\alias{jaw_vars}
\title{View variables needed for simulation}
\usage{
jaw_vars()
}
\description{
View variables needed for simulation and their description
L
}
\details{
This function produces a \code{data.frame} of the names and descriptions of the variables needed for \code{start_sim}. Users of \code{JawCloseR} should reference this before inputing data to \code{start_sim}.
}
\seealso{
\code{\link{start_sime}}
}
