% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{plot_open}
\alias{plot_open}
\title{Plot a start of jaw closing simulation}
\usage{
plot_open(x)
}
\arguments{
\item{x}{a \code{data.frame} a list returned by \code{start_sim()}}
}
\value{
a ggplot object
}
\description{
Plots the output of important positions established by \code{start_sim}.
}
\details{
This function returns a plot of the origin and insertions positions as well as the jaw based on the output of data passed through \code{start_sim}.
}
\examples{
start_sim() \%>\%
   plot_open()

}
