% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{review_data}
\alias{review_data}
\title{Review input data}
\usage{
review_data(x, missing = FALSE)
}
\arguments{
\item{x}{a \code{data.frame} a user intends to pass to \code{start_sim()}}

\item{missing}{logical, should missing variables be returned}
}
\value{
If \code{missing=TRUE}, missing variables are returned as vector.
}
\description{
Review input data, making sure it is appropriate for \code{start_sim}.
}
\details{
This functions tests a \code{data.frame} a user intends to pass to \code{start_sim()}, evaluating if the required variables are present and indicating if the data pass or what variables are missing.
}
\examples{
library(tidyverse)
dat <- system.file("extdata", "csloani.csv", package = "JawCloseR") \%>\% read.csv()
dat <- dat[-4]
miss_vars <- dat \%>\% review_data(missing=TRUE)
print(miss_vars)

}
