% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{rotate_coords}
\alias{rotate_coords}
\title{Plot and print descriptions of configurations supported by \code{JawCloseR}}
\usage{
rotate_coords(xy, rot, center)
}
\arguments{
\item{rot}{numeric, the rotation angle in radians}

\item{center}{numeric, a vector of x and y, the orig about which to rotate the points}

\item{config}{\code{data.frame} of x and y coordinates in the first and second columns, respectively, or a 1D vector of x and y in that order.}
}
\value{
a \code{data.frame} or vector, depending on what is passed to \code{xy}
}
\description{
Plots approximate positions of muscle divisions in the different configurations supported by JawCloseR.  Prints a description of each muscle division to console.
}
