% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{start_simOLD}
\alias{start_simOLD}
\title{Initiate jaw-closing simulations}
\usage{
start_simOLD(x = NULL, spec.n = 1, config = 1, pars = NULL)
}
\arguments{
\item{x}{a \code{data.frame}, see \code{details}.}

\item{spec.n}{numeric, which of the specimens in the data file should be analyzed.}

\item{config}{numeric, the muscle configuration. See \code{details}.}

\item{pars, }{a list of named numerical values. If \code{NULL}, the default values will be used. See \code{details}.}
}
\description{
Establishes starting values for important morphological and physiological parameters to be used in adduction simulations with \code{close_jaw()}.
}
\details{
This function takes morphometric data of the feeding system and prepares it for analysis with \code{close_jaw}. It essentially uses the input data to open the jaw to a specified or default gape angle and outputs important muscle and model parameters to be used by \code{close_jaw()}.

\code{x} should be a \code{data.frame} containing only numeric values with variables/column names specified by \code{jaw_vars()}. The function \code{review_data()} will help new users test data before running \code{close_jaw()}.

\code{config} should be 1, 2, or 3, each corresponding to a specific muscle configuration for simulation according to \insertCite{kenaley2019;textual}{JawCloseR}:

\itemize{
\item \code{config=1}: The \eqn{A_\omega} mass is added to the \eqn{A_2} mass for physiological cross-sectional area calculations. This simplifies simulations, ignoring the geometry of \eqn{A_\omega}, but still accounting for it's force contribution.
\item \code{config=2}: The \eqn{A_\omega} mass and geometry are completely ignored.
\item \code{config=3}: The \eqn{A_\omega} mass and geometry are accounted for as separate muscle units. This is most biologically relevant, of course, but complicates the model considerably.
}
}
\examples{
#Using JawCloseR data from Kenaley et al. (2019) for Chauliodus sloani
cs <- start_sim()
print(cs)

#Using the same, but through entry of a data.frame and the spec.n option
cs_dat<-read.csv(system.file("extdata", "csloani.csv", package = "JawCloseR"))
print(cs_dat)
cs_start <- start_sim(cs_dat,spec.n=2)
print(cs_start)

#Works with pipe (\%>\%), too
library(tidyverse)

cs_dat \%>\%
   start_sim(spec.n=2) \%>\%
   print()
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{close_jaw()}} \code{\link{jaw_vars()}} \code{\link{review_data()}}
}
