% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{var_df}
\alias{var_df}
\title{Establish a simulation data.frame}
\usage{
var_df(x)
}
\arguments{
\item{x}{a list obtained by \code{start_sim}}
}
\description{
Internal function used to establish  a \code{data.frame} where results are stored by \code{close_jaw}.
}
